Cache control buffer
Implements a cache for the control buffer by not reading the control
report for every control value. The cache is used by calling
aqc_get_control_val() and setting hte refresh parameter to false.

This caching feature reduces the amount of requests to the device.
Especially aquaero needed eight requests when using 'sensors' which is
now reduced to one.

Signed-off-by: Leonard Anderweit <leonard.anderweit@gmail.com>